refactor: Move get latest package version implementation to PackageInfo
