+++ src/NPM.ts
@@ -1,11 +1,9 @@
 import { exec } from "child_process"
 
-import { coerce, gt, maxSatisfying } from "semver"
 import { workspace } from "vscode"
 
 import { Cache } from "./Cache"
-import { PackageInfo } from "./PackageInfo"
-import { getCacheLifetime, hasMajorUpdateProtection } from "./Settings"
+import { getCacheLifetime } from "./Settings"
 
 // The `npm view` cache.
 const CACHE_PACKAGES = new Map<string, Cache<Promise<string[]>>>()
@@ -46,53 +44,6 @@ export const getPackageVersions = async (name: string): Promise<string[]> => {
   return execPromise
 }
 
-// Get latest package available, respecting the major update protection, if need.
-export const getPackageLatestVersion = async (
-  packageInfo: PackageInfo
-): Promise<string | null> => {
-  const packageVersions = await packageInfo.getVersions()
-  const versionClean = packageInfo.getVersionClear()
-  const isPrerelease = packageInfo.isVersionPrerelease()
-
-  // We captured the largest version currently available.
-  const versionLatest = maxSatisfying(packageVersions, ">=0", {
-    includePrerelease: isPrerelease,
-  })
-
-  // If protection is not enabled, we will return the latest available version, even if there is a major bump.
-  // Otherwise, we will try to respect the user-defined version limit.
-  if (!hasMajorUpdateProtection()) {
-    return versionLatest
-  }
-
-  // If we are dealing with a user-defined pre-release, we should check the latest compatible non-pre-release version available.
-  // If this version is superior to the current pre-release version, we will suggest it first.
-  if (isPrerelease) {
-    const versionNonPrerelease = maxSatisfying(
-      packageVersions,
-      `^${coerce(versionClean)}`
-    )
-
-    if (versionNonPrerelease && gt(versionNonPrerelease, versionClean)) {
-      return versionNonPrerelease
-    }
-  }
-
-  const versionSatisfying = maxSatisfying(packageVersions, `^${versionClean}`, {
-    includePrerelease: isPrerelease,
-  })
-
-  // If the user-defined version is exactly the same version available within the range given by the user,
-  // we may suggest the latest version, which may include a major bump.
-  // Eg. { "package": "^5.1.3" } and latest is also "5.1.3".
-  if (!versionSatisfying || versionClean === versionSatisfying) {
-    return versionLatest
-  }
-
-  // Otherwise, we will suggest the latest version within the user's range first.
-  return versionSatisfying
-}
-
 interface NPMListResponse {
   dependencies?: Record<string, { version: string }>
 }
@@ -114,7 +65,7 @@ export const getPackagesInstalled = (): Promise<
   const execPromise = new Promise<PackagesInstalled | undefined>((resolve) => {
     const cwd = workspace.workspaceFolders?.[0].uri.fsPath
 
-    return exec(`npm ls --json --depth=0`, { cwd }, (_error, stdout) => {
+    return exec("npm ls --json --depth=0", { cwd }, (_error, stdout) => {
       if (stdout) {
         try {
           const execResult = JSON.parse(stdout) as NPMListResponse
